Add frontend user login and logout (#36)
* User authentication implemented on the frontend

A user that has not logged in will be shown a 'landing' page with a
login button presented. Once logged in the user will view a home page
with a logout button, and the header will populate with links (home and
the admin page are the only links right now).

* Remove "second" page

* API endpoints are now prefixed with `/api`

This allows the UI to reference the API with relative paths, so the UI
does not need to know where the API is hosted. This prevents the UI from
needing env specific configuration.

Note the OAuth callback is not easily changed and backend code will need
to be updated once the OAuth callback is updated.

* Add TTA_SMART_HUB_URI env variable

The TTA_SMART_HUB_URI will replace REDIRECT_URI_HOST as soon as the
localhost OAuth application is updated

* Fix tests

Co-authored-by: kryswisnaskas <krystyna.wisnaskas@gmail.com>