Add the stepper
The stepper accepts any number of "steps" defined as a label an one or
more pages. The stepper can go through these steps in order, backwards
or forwards. If a step has multiple pages The pager is used to go
through each individual page. The main difference between a page and
step are steps are displayed in a "navigation" section at the top of the
stepper, pages do not show up in the stepper "navigation". A simple use
of the stepper is added in the ActivityReport component. This component
will soon be updated/removed once work begins on the actual activity
report.

Some exploritory work has been done with how to use the stepper with
actual form input, which is why `react-hook-form` is used by the stepper
and the ActivityReport component. There is opportunity to make building
forms easier using the stepper and `react-hook-forms` that will be
explored once we impletement a real form.