Update API doc, import specific lodash function, update navigator
The navigator now does not control form data. Form data is kept by the
activity report page. This simplifies passing of data to the review
page. We no longer have to treat the status of the report as a separate
special field, it is simply a field in the report. This change also
helps us make sure that the "managerNotes" field stays updated when a
manager approves a report.