Fix warnings in tests
Most fixes were moving rendering outside of "beforeAll/Each" methods. If
a component updates in a "before*" function after exiting an act warning
is displayed. If render still occurs in a "before*" there is now a line
that waits for rendering to finish

Other various prop checking fixes and waiiting to exit tests until
components update.