Reports can be filtered
The bottom alert table can be filtered. All filtering conditions are
"AND"'ed together. There are two classes of filters, text and dates.

Text filters have "contains" and "does not contain" conditions and
partially matches aginst the specified column. Arrays are searched
through as well by concating all values into one long string.

Date filters have three conditions "Is before", "Is After", and "Is
within". The date picker can be finiky because it wants to be absolute
positioned inside the scrolling container. When opened the date picker
just increases the width of the scrolling container instead of being
placed on top (z-index) of the table. I'm going to work on getting the
table to function properly in a separate PR.

Filters are applied in the backend via "scopes". I didn't use sequelizes
built in scopes because they were not flexible enough. Duplicate columns
between multiple scopes are replaced with sequelize scopes (even duplicate
operators like [Op.or] are replaced). Each filter is convered into a
sequelize literal and everything is tied together inside an [Op.and].