CDI grants can be assigned a grantee and region
The Admin UI allows CDI grants to be assigned a region and grantee. When
a CDI grant is encountered while importing a cdi flag is set to true
(and the region/grantee are not updated). The admin UI looks for this
flag to allow the grant to be managed. After being assigned a grantee
and region the grant can be selected in an activity report.