Context menu can open up instead of down
There are CSS issues overflowing in one axis while scrolling in a different
axis. If one axis scrolls the other axis is set to auto. The result is
context menus get cut off if at the bottom of a scrolling element. This
situation happens with the activity reports table. Added the ability for
the context menu to open up. The last element of the activity report
list now has a usable context menu.