GetWidgets API call handles errors
Unhandled exceptions in handlers never return, which causes a long wait
until the request times out. This causes some issues with the frontend.
GetWidgets now handles errors. Also the totalHrs widget handles the
cause were a time frame is not specified

Async errors are handled automagically in express 5 but not in 4,
https://expressjs.com/en/guide/error-handling.html. We should think
about upgrading to express 5 so we don't run into this situation again
if we forget to manually handle errors in a handler.