Rename grantee to recipient
Update grantee to recipient on the frontend. Grantee TTA record page
hasn't been released, so routes and variables can be changed without
regard to compatability between different versions of the fe and be.
Other places we aren't so lucky. A separate change which will take much
longer will be coming down the pipe to address the grantee -> recipient
change fully in the codebase, including updating routes/the database.