Objectives are saved without explicit button press
Objectives will now be saved as the other fields are, without the need
to explicitly click the "save objective" button. "edit" mode hasn't been
removed, if a user clicks "cancel" they will get the value of the
objective before they started editing (when the edit mode first was
activated). Auto saves, save draft, save and continue will now save the
objective as it is being editing.