Set UTF-8 BOM on CSV files.
Lots of discussion of BOM and if it is good to add to CSV files. See

https://stackoverflow.com/questions/2223882/whats-the-difference-between-utf-8-and-utf-8-without-bom
and
https://stackoverflow.com/questions/17879198/adding-utf-8-bom-to-string-blob/27975629#27975629

The quick explanation is excel doesn't auto detect that our files are
UTF-8 without the BOM set. We set the BOM for UTF-16 which node
translates to the UTF-8 BOM (didn't work when I tried to directly set
the UTF-8 BOM). See the second link above for a possible explanation.
BOM can mess with scripts and combining files together, but I don't
think there is going to be direct manipulation of our CSV files so I
think we are okay setting the BOM.