Goal entry allows pasting from right click
The input was 2px wide set from react-select. This change makes the
input width 100%. Had to move the placeholder because it was stacking
under the input (so the placeholder was half outside of the input box).
Had to set the color of placeholder for accessibility reasons. Class
name is set because I'm not sure the signature of the `cx` prop (if
it can take additional classes or not) so figured the safest bet is
to assume `cx` returns a string (which is required by className) and
just jam everything together in an interpolated string.