Goal ID from API used instead of generated UUID
When saving a goal the frontend will now use returned goals instead of
the object that was just created in the UI. This ensures the ID the UI
sends to the API in subsequent updates matches what is in the database.
This prevents the freshly saved goal from being deleted and recreated
after it has already been created.