Sequelize uses transactions automattically for all queries
See https://sequelize.org/docs/v6/other-topics/transactions/#automatically-pass-transactions-to-all-queries
We no longer need to explicitly pass transactions to each query.
Sequelize uses `cls-hooked` to pass transactions for us. We still have
to kick off the transaction which is handled for all routes with
`transactionWrapper`.

Had to add `waitForExpect` to wait on some transactions to end. For
files it appears supertest gives control back to tests once the `res` is
set in a handler. The file handler does additional processing after
setting the result so the transaction does not finish before tests
regain control, causing the quries inside that request to not be visible
yet (because the transaction hasn't been commited).